+++ test/projects/standard/koot.config.spa.js
@@ -3,4 +3,5 @@ const baseConfig = require('./koot.config')
 module.exports = Object.assign({}, baseConfig, {
     dist: './dist-spa/',
     type: 'react-spa',
+    // templateInject: undefined
 })